Setup data snapshot on page load for the raffle
On load the page will take all firebase data and store it local. So new
donations or changes on the donations will not mess up the tickets. To
fetch the latest data click on the left top of the screen and reset all
local storage. On the next refresh the page will fetch new data and
store it again